fat find: encoding fixes
Use binary IO and byte strings in more places. Some formatting is more
cumbersome, but we really want to use raw bytes when interacting with Git
because that's what it provides us and that's what it expects. Decode when
writing to stdout because that is expected to be read by a human.