Enable git-fat to act like system binary
setup.py follows symlinks and will copy the entire script to env's bin
The previous commit was a workaround to get git-fat to act like a system
binary even when isolated in virtual environments. It was a good
solution if git-fat was more than one file. This solution reverts that
change and should suit our needs better.