boto storage support and a generic storage interface
Support storing git-fat files using boto (which supports S3 and Google
cloud storage.)

The robustness guarantees are identical to the current ones (e.g. if a
git-fat object already exists on either side, it's not checked /
updated, and only a git fat fsck will catch that for you.)

Degrade gracefully if boto cannot be loaded, to allow the original
dependency footprint to continue working.